requirements.txt
numpy>=1.24
pandas>=1.5
matplotlib>=3.7
scipy>=1.10
scikit-learn>=1.3
joblib>=1.3
Pillow>=9.5
flask